backend/requirements.txt
Flask
PyOTP
mysql-connector-python
bcrypt
